Create vbsfdatamover.sh
In case you need to perform a VBSF PoC for a customer who has a Salesforce Org with a lot of data, it is necessary that the Linux server to be used has local space both for backing up data to PostgreSQL (which will occupy 1.6 times the original space) and for metadata and files. 
The customer (which will typically rely on a Linux VM template with a rather small OS partition) will then have to provide an additional partition of sufficient size.
If you have already installed VBSF and PostgreSQL, you can relocate the PostgreSQL 13 data folder to elsewhere by using this script.